+++ glue_vispy_viewers/scatter/layer_style_widget.py
@@ -11,6 +11,7 @@ from glue.external.qt import QtGui
 from glue.utils.qt import load_ui, update_combobox, connect_color
 from glue.utils.qt.widget_properties import (ValueProperty,
                                              CurrentComboProperty,
+                                             CurrentComboTextProperty,
                                              FloatLineProperty,
                                              connect_float_edit,
                                              connect_current_combo,
@@ -21,6 +22,7 @@ from glue.utils.qt.widget_properties import (ValueProperty,
 class ScatterLayerStyleWidget(QtGui.QWidget):
 
     # Size-related GUI elements
+    size_mode = CurrentComboTextProperty('ui.combo_size_mode')
     size_attribute = CurrentComboProperty('ui.combo_size_attribute')
     size_vmin = FloatLineProperty('ui.value_size_vmin')
     size_vmax = FloatLineProperty('ui.value_size_vmax')
@@ -28,6 +30,7 @@ class ScatterLayerStyleWidget(QtGui.QWidget):
     size_scaling = ValueProperty('ui.slider_size_scaling')
 
     # Color-related GUI elements
+    color_mode = CurrentComboTextProperty('ui.combo_color_mode')
     cmap_attribute = CurrentComboProperty('ui.combo_cmap_attribute')
     cmap_vmin = FloatLineProperty('ui.value_cmap_vmin')
     cmap_vmax = FloatLineProperty('ui.value_cmap_vmax')
@@ -53,15 +56,18 @@ class ScatterLayerStyleWidget(QtGui.QWidget):
         self.layer_artist.size = self.layer.style.markersize
         self.layer_artist.size_scaling = 1
         self.layer_artist.size_mode = 'fixed'
-        self.ui.radio_size_fixed.setChecked(True)
+        self.size_mode = 'Fixed'
         self.layer_artist.color = self.layer.style.color
         self.layer_artist.alpha = self.layer.style.alpha
         self.layer_artist.color_mode = 'fixed'
-        self.ui.radio_color_fixed.setChecked(True)
+        self.color_mode = 'Fixed'
         self.ui.combo_size_attribute.setCurrentIndex(0)
         self.ui.combo_cmap_attribute.setCurrentIndex(0)
         self.ui.combo_cmap.setCurrentIndex(0)
         self.layer_artist.visible = True
+
+        self._update_size_mode()
+        self._update_color_mode()
 
     def _connect_global(self):
         connect_float_edit(self.layer.style, 'markersize', self.ui.value_fixed_size)
@@ -73,11 +79,6 @@ class ScatterLayerStyleWidget(QtGui.QWidget):
         pass
 
     def _setup_size_options(self):
-
-        # Set up radio buttons for size mode selection
-        self._radio_size = QtGui.QButtonGroup()
-        self._radio_size.addButton(self.ui.radio_size_fixed)
-        self._radio_size.addButton(self.ui.radio_size_linear)
 
         # Set up attribute list
         label_data = [(comp.label, comp) for comp in self.visible_components]
@@ -91,16 +92,23 @@ class ScatterLayerStyleWidget(QtGui.QWidget):
         connect_value(self.layer_artist, 'size_scaling', self.ui.slider_size_scaling, value_range=(0.1, 10), log=True)
 
         # Set up internal connections
-        self.ui.radio_size_fixed.toggled.connect(self._update_size_mode)
-        self.ui.radio_size_linear.toggled.connect(self._update_size_mode)
+        self.ui.combo_size_mode.currentIndexChanged.connect(self._update_size_mode)
         self.ui.combo_size_attribute.currentIndexChanged.connect(self._update_size_limits)
         self.ui.button_flip_size.clicked.connect(self._flip_size)
 
     def _update_size_mode(self):
-        if self.ui.radio_size_fixed.isChecked():
-            self.layer_artist.size_mode = 'fixed'
+
+        self.layer_artist.size_mode = self.size_mode.lower()
+
+        if self.size_mode == "Fixed":
+            self.ui.size_row_2.hide()
+            self.ui.combo_size_attribute.hide()
+            self.ui.value_fixed_size.show()
         else:
-            self.layer_artist.size_mode = 'linear'
+            self.ui.value_fixed_size.hide()
+            self.ui.combo_size_attribute.show()
+            self.ui.size_row_2.show()
+
 
     def _update_size_limits(self):
 
@@ -118,11 +126,6 @@ class ScatterLayerStyleWidget(QtGui.QWidget):
 
     def _setup_color_options(self):
 
-        # Set up radio buttons for color mode selection
-        self._radio_color = QtGui.QButtonGroup()
-        self._radio_color.addButton(self.ui.radio_color_fixed)
-        self._radio_color.addButton(self.ui.radio_color_linear)
-
         # Set up attribute list
         label_data = [(comp.label, comp) for comp in self.visible_components]
         update_combobox(self.ui.combo_cmap_attribute, label_data)
@@ -136,16 +139,26 @@ class ScatterLayerStyleWidget(QtGui.QWidget):
         connect_value(self.layer_artist, 'alpha', self.ui.slider_alpha, value_range=(0, 1))
 
         # Set up internal connections
-        self.ui.radio_color_fixed.toggled.connect(self._update_color_mode)
-        self.ui.radio_color_linear.toggled.connect(self._update_color_mode)
+        self.ui.combo_color_mode.currentIndexChanged.connect(self._update_color_mode)
         self.ui.combo_cmap_attribute.currentIndexChanged.connect(self._update_cmap_limits)
         self.ui.button_flip_cmap.clicked.connect(self._flip_cmap)
 
     def _update_color_mode(self):
-        if self.ui.radio_color_fixed.isChecked():
-            self.layer_artist.color_mode = 'fixed'
+
+        self.layer_artist.color_mode = self.color_mode.lower()
+
+        if self.color_mode == "Fixed":
+            self.ui.color_row_2.hide()
+            self.ui.color_row_3.hide()
+            self.ui.combo_cmap_attribute.hide()
+            self.ui.spacer_color_label.show()
+            self.ui.label_color.show()
         else:
-            self.layer_artist.color_mode = 'linear'
+            self.ui.label_color.hide()
+            self.ui.combo_cmap_attribute.show()
+            self.ui.spacer_color_label.hide()
+            self.ui.color_row_2.show()
+            self.ui.color_row_3.show()
 
     def _update_cmap_limits(self):
 
@@ -178,3 +191,53 @@ class ScatterLayerStyleWidget(QtGui.QWidget):
             return self.layer.data.visible_components
         else:
             return self.layer.visible_components
+
+
+if __name__ == "__main__":
+
+    from glue.external.qt import get_qapp
+    from glue.external.echo import CallbackProperty
+
+    app = get_qapp()
+
+    class Style(object):
+        color = CallbackProperty("#334455")
+        alpha = CallbackProperty(1.0)
+        markersize = CallbackProperty(4)
+
+    class Component(object):
+        def __init__(self, label):
+            self.label = label
+
+    c1 = Component('a')
+    c2 = Component('b')
+    c3 = Component('c')
+
+    class Layer(object):
+        style = Style()
+        visible_components = [c1, c2, c3]
+
+    class LayerArtist(object):
+        layer = Layer()
+
+        size_mode = CallbackProperty(None)
+        size = CallbackProperty()
+        size_attribute = CallbackProperty()
+        size_vmin = CallbackProperty()
+        size_vmax = CallbackProperty()
+        size_scaling = CallbackProperty()
+
+        color_mode = CallbackProperty(None)
+        color = CallbackProperty()
+        cmap_attribute = CallbackProperty()
+        cmap_vmin = CallbackProperty()
+        cmap_vmax = CallbackProperty()
+        cmap = CallbackProperty()
+        alpha = CallbackProperty()
+
+    layer_artist = LayerArtist()
+
+    widget = ScatterLayerStyleWidget(layer_artist)
+    widget.show()
+
+    app.exec_()