+++ CHANGES.md
@@ -1,7 +1,10 @@
 0.3 (unreleased)
 ----------------
 
-- No changes yet
+- Workaround OpenGL issue that caused cubes with size > 2048 along any
+  dimension to not display.
+
+- Fix issue with _update_data on base VisPy viewer.
 
 0.2 (2015-03-11)
 ----------------