+++ CHANGES.md
@@ -4,6 +4,8 @@
 - Make sure an error is raised if data is not 3-dimensional and shape doesn’t
   agree with existing data in volume viewer.
 
+- Fix a bug that caused exceptions when clearing/removing layer artists. [#117]
+
 - Workaround OpenGL issue that caused cubes with size > 2048 along any
   dimension to not display.